fix: Switch to monitoring_es type
After learning a bit more about how slingshot generates data, I think this structure might work out better. So I can have one monitoring.json that uses es, kibana, etc generators
